% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCCARS.R
\name{rCCARS}
\alias{rCCARS}
\title{Concave-Convex Adaptive Rejection Sampling Algorithm}
\usage{
rCCARS(n, cvformula, ccformula, min, max, sp)
}
\arguments{
\item{n}{Desired sample size (positive integer).}

\item{cvformula}{Convex part of \code{-log(p(x))}, as a character string.}

\item{ccformula}{Concave part of \code{-log(p(x))}, as a character string.}

\item{min, max}{Bounded domain limits (must be finite).}

\item{sp}{Supporting set; a numeric vector of initial abscissae.}
}
\value{
A numeric vector of length \code{n} containing samples from the target distribution.
}
\description{
rCCARS generates a sequence of random numbers by the concave-convex adaptive
rejection sampling algorithm from target distributions with bounded domain.
}
\details{
Strictly speaking, the concave-convex adaptive rejection sampling
  algorithm can generate samples from target distributions with bounded
  domains. For distributions with unbounded domain, rCCARS can also be used
  for approximate sampling by truncating the domain.
}
\examples{
\dontrun{
set.seed(123)
x <- rCCARS(100, "x+x^-1", "2*log(x)", 0.001, 100, 1)
hist(x, breaks = 20, probability = TRUE)
}

}
\references{
Teh Y W. Concave-Convex Adaptive Rejection Sampling[J]. Journal of Computational & Graphical Statistics, 2011, 20(3):670-691.
}
\author{
Dong Zhang
}
