% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMARS.R
\name{rMARS}
\alias{rMARS}
\title{Modified Adaptive Rejection Sampling Algorithm}
\usage{
rMARS(n, formula, min = -Inf, max = Inf, sp, infp, m = 1e-04)
}
\arguments{
\item{n}{Desired sample size (positive integer).}

\item{formula}{Kernel of the target distribution, as a character string.}

\item{min, max}{Domain of the target distribution, including \code{-Inf} and \code{Inf}.}

\item{sp}{Supporting set; a numeric vector of initial abscissae.}

\item{infp}{Inflexion set; a numeric vector of inflexion points of \code{-log(p(x))}.}

\item{m}{A small numeric parameter for judging concavity and convexity near inflexion points.}
}
\value{
A numeric vector of length \code{n} containing samples from the target distribution.
}
\description{
rMARS generates a sequence of random numbers using the modified adaptive
rejection sampling algorithm, which extends ARS to non-log-concave densities.
}
\examples{
set.seed(123)
x <- rMARS(100, "exp(-(4-x^2)^2)", -Inf, Inf,
           c(-2.5, 0, 2.5), c(-2/sqrt(3), 2/sqrt(3)))
hist(x, probability = TRUE, xlim = c(-3, 3), ylim = c(0, 1.2), breaks = 20)
lines(density(x, bw = 0.05), col = "blue")

}
\references{
Martino L, Miguez J. A generalization of the adaptive rejection sampling algorithm[J]. Statistics & Computing, 2011, 21(4):633-647.
}
\author{
Dong Zhang
}
